Add ruff/mypy/bandit quality gates and make the whole repo pass them
Tooling (mirroring the house style of the sibling repos):
- pyproject.toml: ruff (line-length 120, py311, E/W/F/I/B/UP/C4/SIM/RUF,
  per-file ignores for the tests' sys.path shim and the two Monopoly
  data-table modules), shared mypy flags (check_untyped_defs,
  no_implicit_optional, warn_redundant_casts, warn_unused_ignores), and
  bandit with documented B101/B311/B110 skips. The four game folders
  share module names, so mypy runs once per folder.
- .pre-commit-config.yaml: check-only ruff plus merge-conflict/yaml/
  large-file/debug-statement hooks.
- requirements-dev.txt: exact-pinned ruff/mypy/bandit/pytest/pytest-cov/
  pip-audit/pre-commit toolchain.

Code changes to reach a clean bill:
- ~180 auto-fixes (pep585/604 annotations, import order, pyupgrade) plus
  hand fixes for long lines, collapsible branches, contextlib.suppress,
  zip strict=True, and one unused variable.
- Ludo: TokenState now declares _finished_steps as a real field instead
  of an injected attribute; LudoGame accepts a Sequence of seat specs;
  color helpers return explicit RGB 3-tuples; roll handling narrows the
  die value; UI smoke tests import helpers directly.
- Monopoly: an explicit _require_game() invariant helper replaces
  Optional derefs across the app layer; auction/trade/pending-purchase
  states are narrowed where dialogs guarantee them.
- Orbital Orchard: merge candidates are typed against CelestialBody;
  nebula loop variables no longer shadow a differently-shaped tuple.
- Solitaire: annotations for the hint cache and tkinter label style.

All four test suites, the three headless autotests, ruff, mypy (x4),
bandit, and compileall pass.

+++ Python Monopoly Game/tests/test_engine_regressions.py
@@ -6,7 +6,6 @@ import sys
 import unittest
 from pathlib import Path
 
-
 GAME_DIR = Path(__file__).resolve().parents[1]
 if str(GAME_DIR) not in sys.path:
     sys.path.insert(0, str(GAME_DIR))
@@ -26,7 +25,8 @@ def make_game() -> MonopolyGame:
 def rig_rolls(game: MonopolyGame, *rolls: int) -> None:
     """Force the next dice values so a test can land on chosen spaces."""
     pending = list(rolls)
-    game.rng.randint = lambda _start, _end: pending.pop(0)
+    # Deliberate monkeypatch of the game's RNG; tests own this game object.
+    game.rng.randint = lambda _start, _end: pending.pop(0)  # type: ignore[method-assign, assignment]
 
 
 class EngineRegressionTests(unittest.TestCase):
@@ -80,7 +80,7 @@ class EngineRegressionTests(unittest.TestCase):
         self.assertEqual(
             player.cash,
             cash_before + game.board[5].mortgage
-            - int(round(game.board[5].mortgage * 1.1)),
+            - round(game.board[5].mortgage * 1.1),
         )
 
     def test_save_load_round_trip_keeps_property_state_and_ai_profile(self) -> None:
